==> 2.rateBinning_merged_PlanType_Metal-Use.ipynb <==
## Overview
Merge data between sample population data of 6000 records with Plantype and MatalLevel data to combine into one data for easy analysis.

### Objectives
1. Loading and inspecting dataset
2. Observations
3. Select particular BusinessYear 2014
4. Select only the columns you want to analyze
5. Merge sample population 6000 records with Plan Attribute

### 1. Loading and inspecting dataset
#### import library
1. pandas that it allows you to create, manipulate, and analyze datasets efficiently.
2. numpy that it provides support for arrays, matrices, and various mathematical functions.
3. seaborn that it provides high-level functions to create attractive and informative plots.
4. matplotlib is a foundational library for creating static, interactive, and animated visualizations.

import pandas as pd
import numpy as np
import seaborn as sns
import matplotlib.pyplot as plt

# Load the dataset
df_sr = pd.read_csv('sample_rate.csv', low_memory=False)
df_pa = pd.read_csv('PlanAttributes.csv',low_memory=False)

# inspect the dataset
df_sr.shape

# display all columns
pd.set_option('display.max_columns',None)

# display first 5 rows
df_sr.head()

# display last 5 rows
df_sr.tail()

# print first 5 row all columns
print(df_sr.head())

# display all columns, Non-Null Count and Dtype
df_sr.info()

### 2. Observations
1. **Total raws and colums:** The dataset contains 6,000 rows(entries) and 4 columns
2. **Data Types and possible issues**
    - 1 Columns are int64: **Age**
    - 2 Columns are object: **PlanId**

# Countns BusinessYear
df_pa['BusinessYear'].value_counts().reset_index()

# display all columns
pd.set_option('display.max_columns',None)

# print first 5 row all columns
df_pa.head()

### 3. Select particular BusinessYear 2014
1. Accesses the column BusinessYear from the DataFrame df_pa
3. Counts the occurrences of each unique value in the BusinessYear column
4. Resets the index of the Series so that the unique values and their counts become columns in the new DataFrame

# Keep BusinessYear 2014 in df_pa2014
df_pa2014 = df_pa[df_pa['BusinessYear'] == 2014]

# 18,719 records
df_pa2014['BusinessYear'].value_counts().reset_index()

# Counts values and reset index "PlanType"
df_pa2014['PlanType'].value_counts().reset_index

# Counts values and reset index "MetalLevel"
df_pa2014['MetalLevel'].value_counts().reset_index

### 4. Select only the columns you want to analyze
1. Select 'StandardComponentId','PlanType','MetalLevel' columns
2. Remove dupicate rows in 'StandardComponentId' to new variables

#Select only the columns you want to variable df_plantype_metal
df_plantype_metal = df_pa2014[['StandardComponentId','PlanType','MetalLevel']]

# display first 5 rows
df_plantype_metal.head()

# Remove duplicate rows from a DataFrame, considering only the StandardComponentId column
df_duplicate = df_plantype_metal.drop_duplicates(subset=["StandardComponentId"], keep="first")
df_results = df_duplicate[['StandardComponentId','PlanType','MetalLevel']].reset_index()
df_results.info()

### 5. Merge sample population 6000 records with Plan Attribute
1. Merge dr_sr = sample_rate.csv is left with "PlanId" and df_results = PlanAttributes.csv with "StandardComponentId"
2. Create csv file

# Merge file sample_rate.csv and PlanAttributes.csv
merged_df = pd.merge(df_sr, df_results, left_on='PlanId', right_on='StandardComponentId', how='left')

# Display all columns, Non-Null Count and Dtype
merged_df.info()

# Create CSV file
merged_df.to_csv('sample_merged_plantype.csv')
