==> 3.BenefitCostSharing_clean_mean-Use.ipynb <==
## Overview
Clean BenefitCostSharing by Handle missing value and Sepatate file between numeric, string, % and $ in CoinsInnTier1 column

### Objectives
1. Loading and inspecting dataset
2. Select particular BusinessYear 2014
3. Select only the columns that you want to analyze
4. Separate the string to find groups of string
5. Information used in Two-way ANOVA
6. Handle this data
7. Find out Mean of Coinsurance
8. Drop Duplicates values

### 1. Loading and inspecting dataset
#### import library
1. pandas that it allows you to create, manipulate, and analyze datasets efficiently.
2. numpy that it provides support for arrays, matrices, and various mathematical functions.
3. seaborn that it provides high-level functions to create attractive and informative plots.
4. matplotlib is a foundational library for creating static, interactive, and animated visualizations.

import pandas as pd
import numpy as np
import seaborn as sns
import matplotlib.pyplot as plt

# Load the dataset
df_bcs = pd.read_csv("BenefitsCostSharing.csv",low_memory=False)

# inspect the dataset
df_bcs.shape

# Countns BusinessYear
df_bcs['BusinessYear'].value_counts().reset_index()

### 2. Select particular BusinessYear 2014
1. Accesses the column BusinessYear from the DataFrame df_bcs2014
3. Counts the occurrences of each unique value in the BusinessYear column
4. Resets the index of the Series so that the unique values and their counts become columns in the new DataFrame

# Keep BusinessYear 2014 in df_bcs2014
df_bcs2014 = df_bcs[df_bcs['BusinessYear'] == 2014]

# 1,164,869 records
df_bcs2014['BusinessYear'].value_counts().reset_index()

## display all columns
pd.set_option('display.max_columns',None)
# display first 5 rows
df_bcs2014.head()

# inspect the dataset
df_bcs2014.shape

### 3. Select only the columns that you want to analyze
1. Select 'StandardComponentId','BenefitName','CoinsInnTier1' columns.
2. Check 'CoinsInnTier1' Column for to calculate Coinsurance value.

# Select only the columns you want to variable df_c1
df_c1 = df_bcs2014[['StandardComponentId','BenefitName','CoinsInnTier1']]

## display all columns
pd.set_option('display.max_columns',None)
df_c1.head()

# display all columns, Non-Null Count and Dtype
df_c1.info()

# display first 5 rows
df_c1.head()

## display all rows
pd.set_option('display.max_rows',None)
df_c1['CoinsInnTier1'].value_counts().reset_index()

### 4. Separate the string to find groups of string
1. Separate numbers from CoinsInnTier1 and store it in Coin
2. Remove the number part from CoinsInnTier1 and keep only the description (Description)
3. Clean NaN or empty values.

# Separate into Amount in 'Coin' and 'Description'
df_c1['Coin'] = df_c1['CoinsInnTier1'].str.extract(r'([\d,]+)').replace({',': ''}, regex=True)
# Separate only text (cut out numbers)
df_c1['Description'] = df_c1['CoinsInnTier1'].str.replace(r'[\d,]+', '', regex=True).str.strip()

# display first 5 rows
df_c1.head()

### 5. Information used in Two-way ANOVA
The dependent variable must be quantitative in order to be able to analyze differences in means between groups. Therefore, the data in the example you provide must be adjusted or manipulated to provide a suitable numerical value for analysis.

#### 5.1 Understanding Coinsurance Information
From the information you have:
1. No Charge and No Charge after deductible: indicates that there is no charge. or not waste money
2. NaN: May mean missing or unknown data.
3. For example `10%` Coinsurance after deductible and `10%` : should be converted to numbers that can be analyzed statistically.

#### 5.2 Handle non-numeric values
1. `10%, 20%, 30%`: Numbers are values that should be converted to numbers (0.1, 0.2, or 0.3, etc.).
2. No Charge, No Charge after deductible and $: should be converted to 0.
3. NaN: Show missing information that it can be drop because we don't domain expert.
4. Text formats such as `%` Coinsurance after deductible, `%`, No Charge, No Charge after deductible and `$` should be separated or omitted if they are not relevant to the numbers.

# Check Desciption values
value_counts = df_c1['Description'].value_counts().reset_index(name='Count')
value_counts

# Used to retrieve all unique values from the Description column in df_c1.
df_c1['Description'].unique()

# Used to retrieve all unique values from the Coin column in df_c1.
df_c1['Coin'].unique()

### 6. Handle this data
1. Replace No Charge, No Charge after deductible and $ in Description columns = 0 in Coinsurance columns
2. Drop NaN in Decriptions columns

# Step1: Replace 'No Charge', 'No Charge after deductible', '0%' in Description to Coinsurance is 0
df_c1.loc[df_c1['Description'].isin(['No Charge', 'No Charge after deductible', '$']), 'Coin'] = '0'

# Step2: Drop 'NaN' in Coinsurance columns
# df_c1 = df_c1[df_c1['Description'] != 'NaN'].reset_index(drop=True)
df_cleaned = df_c1.dropna()

# Used to convert data values in columns to numeric type.
df_cleaned['Coin'] = pd.to_numeric(df_cleaned['Coin'], errors='coerce')

# Used to retrieve all unique values from the Coin column in df_c1
df_cleaned['Coin'].unique()

# Used to retrieve all unique values from the Description column in df_c1
df_cleaned['Description'].unique()

# display first 5 rows
df_cleaned.head()

# display all columns, Non-Null Count and Dtype
df_cleaned.info()

### 7. Find out Mean of Coinsurance
1. Find out Mean of Coinsurance each PlanId and Add new columns Mean Coinsurance for check each PlanId have how mean of Coinsurance
2. Check manual
3. Calculate the Mean of each PlanId and take the resulting Mean file and add it as a new Column after the Description column by inserting the reference value from the original PlanId
4. Check Coinsurance of StandardComponentId that you want

# Step1: Find out Copayment > 0 for test find out mean value of PlanId
plans_with_copayment = df_cleaned[df_cleaned['Coin'] > 0]['StandardComponentId'].unique()
plans_with_copayment

# Step2: Check manual
# Filtered PlanId is 31274WV0420031 (Example)
filtered_df = df_cleaned[df_cleaned['StandardComponentId'] == '31274WV0420031']

# Count all value of Coin
coin_count = filtered_df['Coin'].notna().sum()

# Find valure that not duplicate value in Coinsurance and number of value
coin_value_counts = filtered_df['Coin'].value_counts().reset_index()
coin_value_counts.columns = ['Coin', 'Count']

# Calculate Coinsurance
coin_mean = filtered_df['Coin'].mean()

# Display recheck mean value of 31274WV0420031
print(f"all Coinsurance values: {coin_count}")
print("Values in Coin:")
print(coin_value_counts)
print(f"Mean of Coinsurance: {coin_mean:.2f}")

#Check manual 31274WV0420031 = ((20x42)+(50x2))/62 = 15.16 Mean

# Step3: Calculate Mean of Copayment for each PlanId
mean_coin_per_plan = df_cleaned.groupby('StandardComponentId')['Coin'].mean().reset_index(name='Coinsurance')

# Merge the Mean value back into the original DataFrame
df_with_mean = pd.merge(df_cleaned, mean_coin_per_plan, on='StandardComponentId', how='left')

# Reorder columns to be after Description
columns_order = df_cleaned.columns.tolist() + ['Coinsurance']
df_with_mean = df_with_mean[columns_order]

# Display first 5 rows
df_with_mean.head()

# display all columns, Non-Null Count and Dtype
df_with_mean.info()

# Step4: Check Copayment of PlanId that you want
mean_value = df_with_mean.loc[df_with_mean['StandardComponentId'] == '31274WV0420031', 'Coinsurance'].iloc[0]

# Display
print(f"Coinsurance value of PlanId 31274WV0420031 is: {mean_value:.2f}")

# Filter Particular PlanId that you want "31274WV0420031"
filtered_df = df_with_mean.loc[df_with_mean['StandardComponentId'] == '31274WV0420031']

# Display all rows
filtered_df

# Rounds a number up to the next integer value.
df_with_mean['Coinsurance'] = np.ceil(df_with_mean['Coinsurance']).astype(int)

# Filter Particular PlanId that you want "31274WV0420031"
filtered_df = df_with_mean.loc[df_with_mean['StandardComponentId'] == '31274WV0420031']

# Display all rows
filtered_df

### 8. Drop Duplicates values
1. Drop 'StandardComponentId' column with duplicate values
2. Create CSV file

#drop duplicates values keep in df_no_duplicates
df_no_duplicates = df_with_mean.drop_duplicates(subset=["StandardComponentId"])

# display all columns, Non-Null Count and Dtype
df_no_duplicates.info()

# display first 5 rows
df_no_duplicates.head()

# Used to retrieve all unique values from the Coin column in df_no_duplicates
df_no_duplicates['Coinsurance'].unique()

# Create csv after clean BenefitCostSharing no duplicates
df_no_duplicates.to_csv('clean_benefitcostsharing_mean_no_duplicates.csv')
