==> 4.Final_Merge_Benefit_Left-Use.ipynb <==
## Overview
Merge data between sample population data of 6000 records, Plantype and MatalLevel in PlanAttributes, Benefitname and Coinsurance in BenefitCostSharing to combine into one data for easy analysis is [AgeGroup, PlanType, Coinsurance]

### Objectives
1. Loading and inspecting dataset
2. Merge sample population 6000 records, PlanType with Coinsurance
3. Select Only Column that use to analyze

### 1. Loading and inspecting dataset
#### import library
1. pandas that it allows you to create, manipulate, and analyze datasets efficiently.
2. numpy that it provides support for arrays, matrices, and various mathematical functions.
3. seaborn that it provides high-level functions to create attractive and informative plots.
4. matplotlib is a foundational library for creating static, interactive, and animated visualizations.

import pandas as pd
import numpy as np
import seaborn as sns
import matplotlib.pyplot as plt

# Load the dataset after cleaning from 3.BenefitCostSharing_clean_mean-Use.ipynb
df_bm = pd.read_csv("clean_benefitcostsharing_mean_no_duplicates.csv",low_memory=False)

# display all columns, Non-Null Count and Dtype
df_bm.info()

# Check NaN value sum
df_bm.isna().sum()

# Load the dataset sample after merged plantype from 2.rateBinning_merged_PlanType_Metal-Use
df_sm = pd.read_csv("sample_merged_plantype.csv",low_memory=False)

# display all columns, Non-Null Count and Dtype
df_sm.info()

# Check NaN value sum
df_sm.isna().sum()

### 2. Merge sample population 6000 records, PlanType with Coinsurance
1. Merge dr_sm = sample_merged_plantype.csv and "StandardComponentId" clean_benefitcostsharing_mean_no_duplicates.csv with "StandardComponentId"
2. Check NaN value
3. Fill 0 value replaced with NaN value
4. Drop unnecessary columns

# Step1: Merge
# Perform the merge
merged_df = pd.merge(df_sm, df_bm, on='StandardComponentId', how='left')

# Check the number of rows
print(merged_df.shape)

# Step2: Check NaN value sum
merged_df.isnull().sum()

merged_df.info()

# display first 5 rows
merged_df.head()

# Step4: Drop unnecessary columns
df_merge = merged_df.drop(columns=['Unnamed: 0.1', 'Unnamed: 0_x', 'index', 'Unnamed: 0_y'])

# display first 5 rows
df_merge.head()

# display all columns, Non-Null Count and Dtype
df_merge.info()

### 3. Select Only Column that use to analyze
1. Select only AgeGroup, Age, PlanType ,and Coinsurance columns
2. Create CSV file

# Select only AgeGroup, MetalLevel ,and Coinsurance columns
df_final_clened = df_merge[['StandardComponentId', 'Age', 'PlanType','Coinsurance']].reset_index(drop=True)
df_final_clened.head()

# display all columns, Non-Null Count and Dtype
df_final_clened.info()

#create csv file merge all value that use analyze is complete
df_final_clened.to_csv('cleaned.csv')
