==> 5.Binning-Use.ipynb <==
## Overview
Binning "Age" column is Common age ranges for health-related analysis include the following categories. By Fixed Interval Binning, a simple method to create age groups. Analyze different age groups and plan types together affect the dependent variable (Coinsurance)

### Objective
1. Loading and inspecting dataset
2. Binning "Age" column
3. Observations
4. Handle unnecessary columns

### 1. Loading and inspecting dataset
#### import library
1. pandas that it allows you to create, manipulate, and analyze datasets efficiently.
2. numpy that it provides support for arrays, matrices, and various mathematical functions.
3. seaborn that it provides high-level functions to create attractive and informative plots.
4. matplotlib is a foundational library for creating static, interactive, and animated visualizations.

import pandas as pd
import numpy as np
import seaborn as sns
import matplotlib.pyplot as plt

# Load the dataset csv file merge all value that use analyze from 4.Final_Merge_Benefit_Left-Use.ipynb
df = pd.read_csv("cleaned.csv",low_memory=False)

# display all columns, Non-Null Count and Dtype
df.info()

# Sort Age
sorted_ages = np.sort(df['Age'].unique())
print(sorted_ages)

# display first 5 rows
df.head()

### 2. Binning "Age" column
1. Common age ranges for health-related analysis include the following categories
2. By Fixed Interval Binning, a simple method to create age groups
3. Analyze different age groups and plan types together affect the dependent variable (Coinsurance) 
    - 0-20: Typically, those who are minors and covered under parents' or guardians' health insurance plans.
    - 21-25: Young adults, often covered under parental insurance until 26.
    - 26-40: Typically, adults in early to mid-career stages, where private insurance plans or employer-sponsored plans become more common.
    - 41-60: Pre-retirement age group, often with an increased focus on preventative care and management of chronic conditions.
    - 61-64: Near-retirement age, typically covered under employer-sponsored plans or private health insurance, just before eligibility for Medicare.
    - 65+: Seniors eligible for Medicare
4. Create a new column for age bins

# Common age bins for health-related analysis include the following categories 
# by Fixed Interval Binning, a simple approach to create age bins 
# and to analyze how different age groups and plan types together affect the dependent variable
# 1. 0-20: Typically, those who are minors and covered under parents' or guardians' health insurance plans.
# 2. 21-25: Young adults, often covered under parental insurance until 26.
# 3. 26-40: Typically, adults in early to mid-career stages, where private insurance plans or employer-sponsored plans become more common.
# 4. 41-60: Pre-retirement age group, often with an increased focus on preventative care and management of chronic conditions.
# 5. 61-64: Near-retirement age, typically covered under employer-sponsored plans or private health insurance, just before eligibility for Medicare.
# 6. 65+: Seniors eligible for Medicare

bins = [0,21, 26, 41, 61, 65, float('inf')]  # Define age ranges
labels = ['0-20', '21-25', '26-40', '41-60', '61-64', '65+']  # Corresponding labels

# Create a new column for age bins
df['AgeGroup'] = pd.cut(df['Age'], bins=bins, labels=labels, right=False)
df_final = df[['AgeGroup', 'Age', 'PlanType','Coinsurance']].reset_index().sort_values(by='AgeGroup')
#df_sorted = df_ra_clean[['AgeGroup', 'PlanId']].value_counts().reset_index(name='Count').sort_values(by='AgeGroup')

# Reset index
df_final.reset_index(drop=True)

# display all columns, Non-Null Count and Dtype
df_final.info()

### 3. Observations
1. **Total raws and colums:** The dataset contains 6,000 rows(entries) and 5 columns
2. **Data Types and possible issues**
    - 2 Columns are int64: **Index, Age**
    - 1 Columns are category: **AgeGroup**
    - 1 Columns are object: **PlanType**
    - 1 Columns are float64: **Copayment**

# display first 5 rows
df_final.head()

# Check for unique values in the AgeGroup column
df_final['AgeGroup'].unique()

# Check for unique values in the PlanType column
df_final['PlanType'].unique()

# Check for unique values in the Age column
df_final['Age'].unique()

### 4. Handle unnecessary columns
1. Drop 'index' columns
2. Create CSV file

# Step1: Drop 'index' columns
df_final = df_final.drop(columns=['index'])
# display first 5 rows
df_final.head()

# Set the original column name (Coinsurance) and new name (Coinsurance Percentage)
df_final.rename(columns={'Coinsurance': 'Coinsurance Percentage'}, inplace=True)

# Convert percentage to decimal number
df_final['Coinsurance'] = df['Coinsurance'] / 100

# Drop unuse column
df_final = df_final.dropna(subset=['Coinsurance Percentage'])

# Sort 'Age' column
df_final = df_final.sort_values(by='Age')

# Reset index and drop index columns
df_final = df_final.reset_index(drop=True)

# Display
df_final

# display all columns, Non-Null Count and Dtype
df_final.info()

# Used to retrieve all unique values from the Coin column in df_final
df_final['Coinsurance'].unique()

# Create CSV file
df_final.to_csv('final_binning.csv')
